Initial non-optimised version of my fd list code.
The idea is that for the most part, we can detect busy and idle clients
at IO register time (and, at the outside, we can re-decide when we get
a ping back from a client..)

The FD list is passed through comm_setselect() as its second argument.
There will be some weirdnesses where FDs shift lists frequently, but I
will play with this code soon.

Read include/fdlist.h for the list of FD lists.