Lee is going to bitchslap me for this one but I don't care(and this will probably make the next release 1.4 oh well..)
Remove the remaining traces of CAP_UID and save ourselves some memory in the
process.

Moved the base64 functions that cryptlinks relies on into irc_string.c and
get rid of md5.c/md5.h as we no longer need them.

Merged in the hidden oper and opercheats patches as it seems everyone is at
least using the opercheats patch and the hidden oper stuff is moderately
useful.