Kind of a large commit here so let me break down what I did.
First, the core of the ircd compiles as a shared object or an ar archive 
depending on the platform.  The reason for this is, on some platforms you
have loadable modules with unresolved symbols.  By making the core a shared 
object itself, the modules can be linked against it, thus solving the unresolved
symbol issue.(This is the way gaim does it as well).

Also add additional checks in autoconf for various headers. 

Moved oper_up into modules/m_oper.c (this should be okay I suspect)

Also it should be noted that modules aren't installing at this point, so the tree 
is basically busticated.  I will fix this at some point soon.